Turn directory entries for files into FPRs
FPR = file permission record. The scheme dictates that these are stored
in the directory entries for each file.

There's also some new helper functions in auth.py, but in the future, we
might want to move those elsewhere. Public key encryption and
verification won't actually depend on which user is doing them.